requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.8.0
mysql-connector-python==8.4.0
bcrypt==4.1.3
python-dotenv==1.0.1
reportlab==4.2.2
eventlet==0.36.1
Pillow==10.3.0
WTForms==3.1.2
